third_party/blink/web_tests/external/wpt/html/canvas/element/text/2d.text.draw.fontface.notinpage-expected.txt

This is a testharness.js-based test.
[FAIL] @font-face fonts should work even if they are not used in the page
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
